requirements.txt
apache-airflow==2.9.3
psycopg2-binary>=2.9
pytest==7.4.4
pytest-cov==5.0.0
